+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [v0.4.0] - 2022-01-04
 
 ### Changed
 
@@ -45,3 +45,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [v0.1.1]: https://github.com/Tuetuopay/tourniquet/releases/tag/v0.1.1
 [v0.2.0]: https://github.com/Tuetuopay/tourniquet/releases/tag/v0.2.0
 [v0.3.0]: https://github.com/Tuetuopay/tourniquet/releases/tag/v0.3.0
+[v0.4.0]: https://github.com/Tuetuopay/tourniquet/releases/tag/v0.4.0